refactor(outbound): split state store modules
